requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
Django==1.11.7
django-celery-beat==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.3
ephem==3.7.6.0
freeze==1.0.10
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
jellyfish==0.5.6
kombu==4.1.0
lxml==4.1.1
mongoengine==0.15.0
numpy==1.13.3
pandas==0.21.0
parsel==1.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pymongo==3.5.1
pyOpenSSL==17.4.0
python-dateutil==2.6.1
python-epo-ops-client==2.3.1
pytz==2017.3
queuelib==1.4.2
recordlinkage==0.10.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.4.0
selenium==3.7.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
vine==1.1.4
w3lib==1.18.0
xlrd==1.1.0
xmltodict==0.11.0
xvfbwrapper==0.2.9
zope.interface==4.4.3
